Feature/adds swerve odometry

+++ 01_Libraries/ghost_util/src/angle_util.cpp
@@ -1,8 +1,9 @@
 #include "ghost_util/angle_util.hpp"
+#include "ghost_util/unit_conversion_utils.hpp"
 
 namespace ghost_util {
 
-float WrapAngle360(float angle){
+double WrapAngle360(double angle){
 	angle = fmod(angle, 360.0);
 
 	if(angle < 0.0){
@@ -12,7 +13,7 @@ float WrapAngle360(float angle){
 	return angle;
 }
 
-float WrapAngle180(float angle){
+double WrapAngle180(double angle){
 	angle = WrapAngle360(angle);
 
 	if(angle > 180.0){
@@ -22,27 +23,27 @@ float WrapAngle180(float angle){
 	return angle;
 }
 
-float WrapAngle2PI(float angle){
+double WrapAngle2PI(double angle){
 	return DEG_TO_RAD * WrapAngle360(angle * RAD_TO_DEG);
 }
 
-float WrapAnglePI(float angle){
+double WrapAnglePI(double angle){
 	return DEG_TO_RAD * WrapAngle180(angle * RAD_TO_DEG);
 }
 
-float FlipAngle180(float angle){
+double FlipAngle180(double angle){
 	return WrapAngle180(angle + 180.0);
 }
 
-float FlipAnglePI(float angle){
+double FlipAnglePI(double angle){
 	return WrapAnglePI(angle + M_PI);
 }
 
-float SmallestAngleDistDeg(float a2, float a1){
+double SmallestAngleDistDeg(double a2, double a1){
 	a1 = WrapAngle360(a1);
 	a2 = WrapAngle360(a2);
 
-	float diff = a2 - a1;
+	double diff = a2 - a1;
 	if(diff > 180.0){
 		diff = -(360.0 - diff);
 	}
@@ -52,7 +53,7 @@ float SmallestAngleDistDeg(float a2, float a1){
 	return diff;
 }
 
-float SmallestAngleDistRad(float a2, float a1){
+double SmallestAngleDistRad(double a2, double a1){
 	return DEG_TO_RAD * SmallestAngleDistDeg(RAD_TO_DEG * a2, RAD_TO_DEG * a1);
 }
